+++ src/cli/dcos_docker/commands/wait.py
@@ -12,11 +12,16 @@ from cli.common.options import (
     superuser_username_option,
     verbosity_option,
 )
-from cli.common.utils import check_cluster_id_exists, set_logging
+from cli.common.utils import (
+    check_cluster_id_exists,
+    set_logging,
+    show_wait_help,
+)
 from dcos_e2e.node import Transport
 
 from ._common import ClusterContainers, existing_cluster_ids
 from ._options import node_transport_option
+from .doctor import doctor
 
 
 @click.command('wait')
@@ -38,7 +43,9 @@ from ._options import node_transport_option
 )
 @node_transport_option
 @verbosity_option
+@click.pass_context
 def wait(
+    ctx: click.core.Context,
     cluster_id: str,
     superuser_username: str,
     superuser_password: str,
@@ -55,19 +62,23 @@ def wait(
     )
     set_logging(verbosity_level=verbose)
     urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
-    message = (
-        'A cluster may take some time to be ready.\n'
-        'The amount of time it takes to start a cluster depends on a variety '
-        'of factors.\n'
-        'If you are concerned that this is hanging, try "dcos-docker doctor" '
-        'to diagnose common issues.'
-    )
-    click.echo(message)
     cluster_containers = ClusterContainers(
         cluster_id=cluster_id,
         transport=transport,
     )
+
     http_checks = not skip_http_checks
+    parent = ctx.parent
+    assert parent is not None
+    doctor_command_name = '{info_name} {doctor_name}'.format(
+        info_name=parent.info_name,
+        doctor_name=doctor.name,
+    )
+    show_wait_help(
+        is_enterprise=cluster_containers.is_enterprise,
+        doctor_command_name=doctor_command_name,
+    )
+
     with click_spinner.spinner():
         if cluster_containers.is_enterprise:
             cluster_containers.cluster.wait_for_dcos_ee(