+++ src/cli/dcos_vagrant/commands/wait.py
@@ -12,9 +12,14 @@ from cli.common.options import (
     superuser_username_option,
     verbosity_option,
 )
-from cli.common.utils import check_cluster_id_exists, set_logging
+from cli.common.utils import (
+    check_cluster_id_exists,
+    set_logging,
+    show_wait_help,
+)
 
 from ._common import ClusterVMs, existing_cluster_ids
+from .doctor import doctor
 
 
 @click.command('wait')
@@ -22,7 +27,9 @@ from ._common import ClusterVMs, existing_cluster_ids
 @superuser_username_option
 @superuser_password_option
 @verbosity_option
+@click.pass_context
 def wait(
+    ctx: click.core.Context,
     cluster_id: str,
     superuser_username: str,
     superuser_password: str,
@@ -37,15 +44,17 @@ def wait(
     )
     set_logging(verbosity_level=verbose)
     urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
-    message = (
-        'A cluster may take some time to be ready.\n'
-        'The amount of time it takes to start a cluster depends on a variety '
-        'of factors.\n'
-        'If you are concerned that this is hanging, try "dcos-vagrant doctor" '
-        'to diagnose common issues.'
-    )
-    click.echo(message)
     cluster_vms = ClusterVMs(cluster_id=cluster_id)
+    parent = ctx.parent
+    assert parent is not None
+    doctor_command_name = '{info_name} {doctor_name}'.format(
+        info_name=parent.info_name,
+        doctor_name=doctor.name,
+    )
+    show_wait_help(
+        is_enterprise=cluster_vms.is_enterprise,
+        doctor_command_name=doctor_command_name,
+    )
     with click_spinner.spinner():
         if cluster_vms.is_enterprise:
             cluster_vms.cluster.wait_for_dcos_ee(