+++ src/dcos_e2e/cluster.py
@@ -165,7 +165,7 @@ class Cluster(ContextDecorator):
         # outside the cluster for the login and Admin Router only rewrites
         # requests to them, the login endpoint does not provide anything.
 
-        # Current solution to guarantee the security CLI login:
+        # Current solution to guarantee the CLI login:
 
         # Try until one can login successfully with a long lived token
         # (dirty hack in dcos-test-utils wait_for_dcos). This is to avoid
@@ -194,6 +194,32 @@ class Cluster(ContextDecorator):
         )
 
         api_session.wait_for_dcos()  # type: ignore
+
+        # Only the first user can log in with SSO, before granting others
+        # access.
+        # Therefore, we delete the user who was created to wait for DC/OS.
+        #
+        # In order to create an API session, we create a user with the
+        # hard coded credentials "CI_CREDENTIALS".
+        # These credentials match a user with the email address
+        # "albert@bekstil.net".
+        email = 'albert@bekstil.net'
+        path = '/dcos/users/{email}'.format(email=email)
+        server_option = (
+            '"zk-1.zk:2181,zk-2.zk:2181,zk-3.zk:2181,zk-4.zk:2181,'
+            'zk-5.zk:2181"'
+        )
+        delete_user_args = [
+            '.',
+            '/opt/mesosphere/environment.export',
+            '&&',
+            'zkCli.sh',
+            '-server',
+            server_option,
+            'delete',
+            path,
+        ]
+        any_master.run(args=delete_user_args, shell=True, log_output_live=True)
 
     def wait_for_dcos_ee(
         self,
@@ -478,7 +504,7 @@ class Cluster(ContextDecorator):
             subprocess.CalledProcessError: If the ``pytest`` command fails.
         """
         args = [
-            'source',
+            '.',
             '/opt/mesosphere/environment.export',
             '&&',
             'cd',